+++ src/commands/refreshTree.ts
@@ -1,0 +1,15 @@
+import BaseCommand from "../common/baseCommand";
+import * as vscode from 'vscode';
+import { TaskTreeDataProvider } from "../taskTreeDataProvider";
+
+export class RefreshTreeCommand extends BaseCommand {
+  private taskTreeDataProvider: TaskTreeDataProvider;
+  constructor(context: vscode.ExtensionContext) {
+    super('refreshTree', context);
+    this.taskTreeDataProvider = TaskTreeDataProvider.getInstance(context);
+  }
+
+  async run(): Promise<void> {
+    this.taskTreeDataProvider.refreshLocal();
+  }
+}