+++ src/providers/gradleTaskProvider.ts
@@ -63,7 +63,7 @@ export class GradleTaskProvider extends BaseTaskProvider implements TaskProvider
             item.description = vscode.workspace.asRelativePath(file);
             item.startLine = i;
 
-            item.command = {
+            item.onSingleClickCommand = {
               command: 'workspaceTasks.openFileAtLine',
               title: 'Open File',
               arguments: [file, item.startLine || 0]